ggml-sycl: fix compiler warnings
- Remove unused variables (type_size_src0, need_tempBuffers, need_dequant, Tc, stride vars)
- Fix format specifier warnings by casting pointers to void* for %p
- Fix sign comparison warning by casting to size_t
- Suppress unused parameter warnings with (void) casts
- Remove unused BF16_SIZE and FLOAT_SIZE constexpr variables

+++ ggml/src/ggml-sycl/ssm_conv.cpp
@@ -99,7 +99,7 @@ void ggml_sycl_ssm_conv(ggml_backend_sycl_context & ctx, ggml_tensor * dst) {
     GGML_ASSERT(src0->nb[0] == ggml_type_size(src0->type));
     GGML_ASSERT(src1->nb[0] == ggml_type_size(src1->type));
 
-    GGML_ASSERT(src0->nb[1] == src0->ne[0] * (int)ggml_type_size(src0->type));
+    GGML_ASSERT(src0->nb[1] == (size_t)src0->ne[0] * (int)ggml_type_size(src0->type));
 
     const int src_stride_inner = ncs;
     const int src_stride_seq   = ncs * d_inner;
